requirements.txt
selenium>=4.18.1
webdriver-manager>=4.0.1
requests>=2.31.0
google-generativeai>=0.3.2
python-dotenv>=1.0.1
mlflow>=2.10.2
crewai>=0.16.1
langchain>=0.1.9
pyyaml>=6.0.1
psycopg2-binary>=2.9.9
psutil>=5.9.8
